backend/requirements.txt
Django==3.2
djangorestframework==3.14.0
psycopg2==2.9.7
redis==4.5.1
celery==5.2.7
cryptography==41.0.3
requests==2.31.0
Pillow==10.0.0
PyJWT==2.7.0
PyNaCl==1.5.0
django-otp==1.2.0
qrcode==7.4.2
gunicorn==20.1.0

# Now for channels:
channels==3.0.5
channels-redis==3.3.1
